requirements.txt
opencv-python
face_recognition
mysql-connector-python
pandas
streamlit
plotly